Cleaning up FQDNNetworPolicy definition
Reusing the existing NetworkPolicyPort struct
instead of creating a new one. This makes the code
simpler.

The only thing is that a NetworkPolicyPort can be
a string (a named port) which doesn't seem to make
a lot of sense in our context, but I'm not sure
I understand that fully.

Also, moving the helper functions for testing
in a separate file and exporting them. We'll
need them to test the controller.

Change-Id: I0cb1d25a2d90744e6710032c40e505dec66cf985