Introducing v1alpha2 API
To support Ingress policies, we need to modify a bit
the API, so we introduce a v1alpha2.
It adds the "ingress" section to the CRD, but also
a policyType like the one used in NetworkPolicies.
The behavior of that policyType makes this a non-backward
compatible change.

This commit only introduces the new API. There is no
actual implementation yet.